--- PhotoWall/Presentation/PhotoWall/Custom Layouts/CustomLayoutProtocol.swift
@@ -1,13 +1,0 @@
-//
-//  CustomLayoutProtocol.swift
-//  photo-wall
-//
-//  Created by Giovani Nascimento Pereira on 20/05/18.
-//  Copyright © 2018 Giovani Nascimento Pereira. All rights reserved.
-//
-
-import UIKit
-
-class CustomLayout: UICollectionViewLayout {
-    weak var delegate: CustomLayoutDelegate!
-}
