+++ PhotoWall/Presentation/PhotoWall/Custom Layouts/TwoLineGridLayout.swift
@@ -14,8 +14,6 @@ class LinedGridLayout: CustomLayout {
     fileprivate var numberOfLines = 2
     fileprivate var cellPadding: CGFloat = 20
     
-    fileprivate var cache = [UICollectionViewLayoutAttributes]()
-    
     fileprivate var contentHeight: CGFloat {
         guard let collectionView = collectionView else {
             return 0
@@ -73,23 +71,6 @@ class LinedGridLayout: CustomLayout {
             
             line = line < (numberOfLines - 1) ? (line + 1) : 0
         }
-    }
-    
-    override func layoutAttributesForElements(in rect: CGRect) -> [UICollectionViewLayoutAttributes]? {
-        
-        var visibleLayoutAttributes = [UICollectionViewLayoutAttributes]()
-        
-        // Loop through the cache and look for items in the rect
-        for attributes in cache {
-            if attributes.frame.intersects(rect) {
-                visibleLayoutAttributes.append(attributes)
-            }
-        }
-        return visibleLayoutAttributes
-    }
-    
-    override func layoutAttributesForItem(at indexPath: IndexPath) -> UICollectionViewLayoutAttributes? {
-        return cache[indexPath.item]
     }
     
 }
